requirements.txt
google-genai==1.49.0
anthropic==0.72.0
mistralai==1.9.11
together==1.5.30
openai==2.7.1
google-cloud-secret-manager>=2.20.0
google-cloud-storage>=2.14.0
python-dotenv>=1.0.0
isort
black
flake8
flake8-bugbear
pydocstyle
pytest
pytest-cov
pytest-xdist
    